moved rails admin fields to seperate files
