+++ app/lib/rails_admin/config/fields/types/citext.rb
@@ -1,0 +1,9 @@
+module RailsAdmin
+  module Config
+    module Fields
+      module Types
+        class Citext < String; end
+      end
+    end
+  end
+end
