+++ app/lib/rails_admin/config/fields/types/pg_array.rb
@@ -1,0 +1,11 @@
+module RailsAdmin
+  module Config
+    module Fields
+      module Types
+        class PgArray < Base
+          register_instance_option(:formatted_value) { value.join ', ' unless value.blank? }
+        end
+      end
+    end
+  end
+end
