--- lib/rails_admin/config/fields/types/citext.rb
@@ -1,13 +1,0 @@
-require 'rails_admin/config/fields/base'
-
-module RailsAdmin
-  module Config
-    module Fields
-      module Types
-        class Citext < RailsAdmin::Config::Fields::Types::String
-          RailsAdmin::Config::Fields::Types.register :citext, self
-        end
-      end
-    end
-  end
-end
